+++ common/json/infinity.go
@@ -10,7 +10,7 @@ package json
 // Adapted from encoding/json/scanner.go.
 
 // stateI is the state after reading `In`.
-func stateIn(s *scanner, c int) int {
+func stateIn(s *scanner, c byte) int {
 	if c == 'f' {
 		s.step = generateState("Infinity", []byte("inity"), stateEndValue)
 		return scanContinue
